+++ static/app/views/insights/pages/backend/am1BackendOverviewPage.tsx
@@ -32,16 +32,21 @@ import {useUserTeams} from 'sentry/utils/useUserTeams';
 import * as ModuleLayout from 'sentry/views/insights/common/components/moduleLayout';
 import {ToolRibbon} from 'sentry/views/insights/common/components/ribbon';
 import {useOnboardingProject} from 'sentry/views/insights/common/queries/useOnboardingProject';
-import {OVERVIEW_PAGE_ALLOWED_OPS as BACKEND_OVERVIEW_PAGE_ALLOWED_OPS} from 'sentry/views/insights/pages/backend/settings';
-import {FrontendHeader} from 'sentry/views/insights/pages/frontend/frontendPageHeader';
+import {BackendHeader} from 'sentry/views/insights/pages/backend/backendPageHeader';
 import {
-  FRONTEND_LANDING_TITLE,
-  FRONTEND_PLATFORMS,
+  BACKEND_LANDING_TITLE,
   OVERVIEW_PAGE_ALLOWED_OPS,
-} from 'sentry/views/insights/pages/frontend/settings';
-import {useOverviewPageTrackPageload} from 'sentry/views/insights/pages/useOverviewPageTrackAnalytics';
+} from 'sentry/views/insights/pages/backend/settings';
 import {
-  generateFrontendOtherPerformanceEventView,
+  FRONTEND_PLATFORMS,
+  OVERVIEW_PAGE_ALLOWED_OPS as FRONTEND_OVERVIEW_PAGE_OPS,
+} from 'sentry/views/insights/pages/frontend/settings';
+import {
+  MOBILE_PLATFORMS,
+  OVERVIEW_PAGE_ALLOWED_OPS as BACKEND_OVERVIEW_PAGE_OPS,
+} from 'sentry/views/insights/pages/mobile/settings';
+import {
+  generateBackendPerformanceEventView,
   USER_MISERY_TOOLTIP,
 } from 'sentry/views/performance/data';
 import {
@@ -57,31 +62,32 @@ import {
   ProjectPerformanceType,
 } from 'sentry/views/performance/utils';
 
-const DURATION_TOOLTIP = tct(
-  'A heuristic measuring when a pageload or navigation completes. Based on whether the initial load of the webpage has become idle. [link:Learn more.]',
+const APDEX_TOOLTIP = tct(
+  'An industry-standard metric used to measure user satisfaction based on your application response times. [link:Learn more.]',
   {
     link: (
-      <ExternalLink href="https://docs.sentry.io/platforms/javascript/tracing/instrumentation/automatic-instrumentation/#idletimeout" />
+      <ExternalLink href="https://docs.sentry.io/product/performance/metrics/#apdex" />
     ),
   }
 );
 
-const FRONTEND_COLUMN_TITLES = [
+const BACKEND_COLUMN_TITLES = [
+  {title: 'http method'},
   {title: 'transaction'},
   {title: 'operation'},
   {title: 'project'},
   {title: 'tpm'},
-  {title: 'p50()', tooltip: DURATION_TOOLTIP},
-  {title: 'p75()', tooltip: DURATION_TOOLTIP},
-  {title: 'p95()', tooltip: DURATION_TOOLTIP},
+  {title: 'p50()'},
+  {title: 'p95()'},
+  {title: 'failure rate'},
+  {title: 'apdex', tooltip: APDEX_TOOLTIP},
   {title: 'users'},
   {title: 'user misery', tooltip: USER_MISERY_TOOLTIP},
 ];
 
-export function OldFrontendOverviewPage() {
-  useOverviewPageTrackPageload();
+// Am1 customers do not have EAP, so we need to use the old frontend overview page for now
+export function Am1BackendOverviewPage() {
   const theme = useTheme();
-
   const organization = useOrganization();
   const location = useLocation();
   const {setPageError} = usePageAlert();
@@ -93,32 +99,32 @@ export function OldFrontendOverviewPage() {
   const {selection} = usePageFilters();
 
   const withStaticFilters = canUseMetricsData(organization);
-  const eventView = generateFrontendOtherPerformanceEventView(
-    location,
-    withStaticFilters
-  );
+  const eventView = generateBackendPerformanceEventView(location, withStaticFilters);
   const searchBarEventView = eventView.clone();
-
-  const sharedProps = {eventView, location, organization, withStaticFilters};
 
   const segmentOp = 'transaction.op';
 
   // TODO - this should come from MetricsField / EAP fields
   eventView.fields = [
     {field: 'team_key_transaction'},
+    {field: 'http.method'},
     {field: 'transaction'},
     {field: segmentOp},
     {field: 'project'},
     {field: 'tpm()'},
-    {field: 'p50(transaction.duration)'},
-    {field: 'p75(transaction.duration)'},
-    {field: 'p95(transaction.duration)'},
+    {field: 'p50()'},
+    {field: 'p95()'},
+    {field: 'failure_rate()'},
+    {field: 'apdex()'},
     {field: 'count_unique(user)'},
     {field: 'count_miserable(user)'},
     {field: 'user_misery()'},
   ].map(field => ({...field, width: COL_WIDTH_UNDEFINED}));
 
-  const doubleChartRowEventView = eventView.clone(); // some of the double chart rows rely on span metrics, so they can't be queried the same way
+  const doubleChartRowEventView = eventView.clone(); // some of the double chart rows rely on span metrics, so they can't be queried with the same tags/filters
+  const disallowedOps = [
+    ...new Set([...FRONTEND_OVERVIEW_PAGE_OPS, ...BACKEND_OVERVIEW_PAGE_OPS]),
+  ];
 
   const selectedFrontendProjects: Project[] = getSelectedProjectList(
     selection.projects,
@@ -127,27 +133,40 @@ export function OldFrontendOverviewPage() {
     Boolean(project?.platform && FRONTEND_PLATFORMS.includes(project.platform))
   );
 
+  const selectedMobileProjects: Project[] = getSelectedProjectList(
+    selection.projects,
+    projects
+  ).filter((project): project is Project =>
+    Boolean(project?.platform && MOBILE_PLATFORMS.includes(project.platform))
+  );
+
   const existingQuery = new MutableSearch(eventView.query);
-  // TODO - this query is getting complicated, once were on EAP, we should consider moving this to the backend
   existingQuery.addOp('(');
-  existingQuery.addDisjunctionFilterValues(segmentOp, OVERVIEW_PAGE_ALLOWED_OPS);
-  // add disjunction filter creates a very long query as it seperates conditions with OR, project ids are numeric with no spaces, so we can use a comma seperated list
-  if (selectedFrontendProjects.length > 0) {
-    existingQuery.addOp('OR');
+  existingQuery.addOp('(');
+  existingQuery.addFilterValues(`!${segmentOp}`, disallowedOps);
+
+  if (selectedFrontendProjects.length > 0 || selectedMobileProjects.length > 0) {
     existingQuery.addFilterValue(
-      'project.id',
-      `[${selectedFrontendProjects.map(({id}) => id).join(',')}]`
+      '!project.id',
+      `[${[
+        ...selectedFrontendProjects.map(project => project.id),
+        ...selectedMobileProjects.map(project => project.id),
+      ]}]`
     );
   }
   existingQuery.addOp(')');
-  existingQuery.addFilterValues(`!${segmentOp}`, BACKEND_OVERVIEW_PAGE_ALLOWED_OPS);
+  existingQuery.addOp('OR');
+  existingQuery.addDisjunctionFilterValues(segmentOp, OVERVIEW_PAGE_ALLOWED_OPS);
+  existingQuery.addOp(')');
+
   eventView.query = existingQuery.formatString();
 
   const showOnboarding = onboardingProject !== undefined;
 
   const doubleChartRowCharts = [
     PerformanceWidgetSetting.SLOW_HTTP_OPS,
-    PerformanceWidgetSetting.SLOW_RESOURCE_OPS,
+    PerformanceWidgetSetting.SLOW_DB_OPS,
+    PerformanceWidgetSetting.MOST_RELATED_ISSUES,
   ];
   const tripleChartRowCharts = filterAllowedChartsMetrics(
     organization,
@@ -159,15 +178,20 @@ export function OldFrontendOverviewPage() {
       PerformanceWidgetSetting.P95_DURATION_AREA,
       PerformanceWidgetSetting.P99_DURATION_AREA,
       PerformanceWidgetSetting.FAILURE_RATE_AREA,
+      PerformanceWidgetSetting.APDEX_AREA,
     ],
     mepSetting
   );
 
   if (organization.features.includes('insights-initial-modules')) {
+    doubleChartRowCharts.unshift(
+      PerformanceWidgetSetting.HIGHEST_CACHE_MISS_RATE_TRANSACTIONS
+    );
     doubleChartRowCharts.unshift(PerformanceWidgetSetting.MOST_TIME_CONSUMING_DOMAINS);
-    doubleChartRowCharts.unshift(PerformanceWidgetSetting.MOST_TIME_CONSUMING_RESOURCES);
-    doubleChartRowCharts.unshift(PerformanceWidgetSetting.HIGHEST_OPPORTUNITY_PAGES);
+    doubleChartRowCharts.unshift(PerformanceWidgetSetting.MOST_TIME_SPENT_DB_QUERIES);
   }
+
+  const sharedProps = {eventView, location, organization, withStaticFilters};
 
   const getFreeTextFromQuery = (query: string) => {
     const conditions = new MutableSearch(query);
@@ -184,7 +208,7 @@ export function OldFrontendOverviewPage() {
   };
 
   function handleSearch(searchQuery: string) {
-    trackAnalytics('performance.domains.frontend.search', {organization});
+    trackAnalytics('performance.domains.backend.search', {organization});
 
     navigate({
       pathname: location.pathname,
@@ -205,7 +229,7 @@ export function OldFrontendOverviewPage() {
       organization={organization}
       renderDisabled={NoAccess}
     >
-      <FrontendHeader headerTitle={FRONTEND_LANDING_TITLE} />
+      <BackendHeader headerTitle={BACKEND_LANDING_TITLE} />
       <Layout.Body>
         <Layout.Main fullWidth>
           <ModuleLayout.Layout>
@@ -237,24 +261,27 @@ export function OldFrontendOverviewPage() {
                 />
               ) : (
                 <PerformanceDisplayProvider
-                  value={{performanceType: ProjectPerformanceType.FRONTEND_OTHER}}
+                  value={{performanceType: ProjectPerformanceType.BACKEND}}
                 >
-                  <DoubleChartRow
-                    allowedCharts={doubleChartRowCharts}
-                    {...sharedProps}
-                    eventView={doubleChartRowEventView}
-                  />
-                  <TripleChartRow allowedCharts={tripleChartRowCharts} {...sharedProps} />
                   <TeamKeyTransactionManager.Provider
                     organization={organization}
                     teams={teams}
                     selectedTeams={['myteams']}
                     selectedProjects={eventView.project.map(String)}
                   >
+                    <DoubleChartRow
+                      allowedCharts={doubleChartRowCharts}
+                      {...sharedProps}
+                      eventView={doubleChartRowEventView}
+                    />
+                    <TripleChartRow
+                      allowedCharts={tripleChartRowCharts}
+                      {...sharedProps}
+                    />
                     <Table
                       theme={theme}
                       projects={projects}
-                      columnTitles={FRONTEND_COLUMN_TITLES}
+                      columnTitles={BACKEND_COLUMN_TITLES}
                       setError={setPageError}
                       {...sharedProps}
                     />
